requirements.txt
beautifulsoup4==4.10.0
python-dotenv==1.0.0
selenium==4.12.0
sib_api_v3_sdk==7.6.0
